Added functionality to game. Now plays until the player or computer wins 5 times.
Storing game state in a local variable (wins/losses)

Disabled buttons for 1.5 seconds after selection, to allow a moment for the
result of the current round to be seen. Results revealed in a previously hidden
div for the same period of time.

When 5 wins or losses are accumulated, the appropriate gameover message is displayed and
only the Reset Records button is enabled. Clicking reset sthe scores and begins a new game.